+++ Models/Availability.cs
@@ -12,7 +12,7 @@ namespace Gride.Models
 	public class Availability
 	{
 		[Key, ForeignKey("EmployeeID")]
-		public uint EmployeeID { get; set; }
+		public long EmployeeID { get; set; }
 		public DateTime Start { get; set; }
 		public DateTime End { get; set; }
 		public bool Prefered { get; set; } = true;
