+++ Models/Binds/EmployeeAndSkillBind.cs
@@ -10,7 +10,7 @@ namespace Gride.Models.Binds
 	public class EmployeeAndSkillBind
 	{
 		[Key, ForeignKey("Employee")]
-		public uint EmployeeID { get; set; }
+		public long EmployeeID { get; set; }
 
 		[Key, ForeignKey("Skill")]
 		public int SkillID { get; set; }
