+++ Models/Employee.cs
@@ -12,7 +12,7 @@ namespace Gride.Models
 	public class Employee
 	{
 		[Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-		public uint EmployeeID { get; set; }
+		public long EmployeeID { get; set; }
 		[Required, StringLength(50)]
 		public string Name{ get; set; }
 		[Required, StringLength(50), Display(Name="Last Name")]
@@ -25,7 +25,7 @@ namespace Gride.Models
 		[Required, StringLength(12), Phone, Display(Name = "Phone Number")]
 		public string PhoneNumber { get; set; }
 		public bool Admin { get; set; } = false;
-		public ulong LoginID { get; set; }
+		public long LoginID { get; set; }
 		public float Experience { get; set; }
 		[RegularExpression(@"(\\\\?([^\\/]*[\\/])*)([^\\/]+)$", ErrorMessage = "Path to ProfileImage is not a valid path")]
 		public string ProfileImage { get; set; } = null;
